moved test files into subdirectory near source code
