+++ workshop-registration/adapters/list_workshops_presenter_pprint.py
@@ -1,6 +1,6 @@
 from pprint import pprint
 
-from app.list_workshops_app import ListWorkshopsPresenter, WorkshopRegistrationSummary
+from app.list_workshops_workflow import ListWorkshopsPresenter, WorkshopRegistrationSummary
 
 
 class PPrintListWorkshopPresenter(ListWorkshopsPresenter):
