+++ workshop-registration/app/__init__.py
@@ -1,5 +1,5 @@
-from .list_workshops_app import AppModel, ListWorkshopsWorkflow, ListWorkshopsPresenter
+from .list_workshops_workflow import AppModel, ListWorkshopsWorkflow, ListWorkshopsPresenter
 from .workshoprepo import WorkshopRepo, WorkshopRecord
 from .registrationrepo import RegistrationRecord, RegistrationRepo
-from .list_registrants_app import ListRegistrantPresenter, ListRegistrantsWorkflow, RegistrantSummary
+from .list_registrants_workflow import ListRegistrantPresenter, ListRegistrantsWorkflow, RegistrantSummary
 from .app import App
