+++ workshop-registration/external/zoom_api/tests/test_get_meetings.py
@@ -1,4 +1,5 @@
-from unittest.mock import Mock
+from unittest.mock import Mock, patch
+
 from external.zoom_api.get_meetings import get_meetings
