+++ workshop-registration/tests/e2e_test.py
